Download high res profile pictures
I also noticed that int is too small for most Facebook id
Long might eventually be too small too, should migrate to
Strings (they're only for URLs) or BigIntegers

+++ src/org/cjones/facebook/Friend.java
@@ -7,13 +7,13 @@ import android.net.Uri;
 
 public class Friend
 {
-    private int id;
+    private long id;
     private String name;
     private byte[] photo_bytes;
     private Bitmap photo;
     private Uri uri;
 
-    public Friend(int id, String name, byte[] photo, Uri uri)
+    public Friend(long id, String name, byte[] photo, Uri uri)
     {
         this.id = id;
         this.name = name;
@@ -27,7 +27,7 @@ public class Friend
         ContactManager.setPhoto(uri, photo_bytes, activity);
     }
 
-    public int getID()
+    public long getID()
     {
         return id;
     }
